At the end of function run_program_and_log_to_screen, paranoid_pclose is used to close the process stream. This means - as we don't evaluate the exit code of the called program - that mondoarchive will happily continue even if mondo has failed. Only later will it give an error when it tries to burn an image that miondi has actually failed to create.
The change to libmondo-fork.c.diff addresses this by using pclose 
directly (rather than the paranoid_pclose macro) and by then evaluating
the status information and extracting the exit code of the called
program.

The change to libmondo-archive.c makes it so that at the end of function 
call_mindi_to_supply_boot_disks popup_and_OK is called in case of an 
error rather than log_to_screen. This ensures that the main error is not 
drowned in a sea of other messages.