WARNING: Not tested yet ! Will be stabilized in next hours. this patch changes the interface between mindi and mondo. Everything is now managed through either MONDO_CACHE/mondo-restore.cfg and MONDO_CACHE/mindi.conf (an additional configuration file) This removes all the one line files NFS-*, ... that where used in an anarchic way. Now everything computed by mondo is under mondo-restore.cfg and everything related to mindi in mindi.conf
Still needed are the removal of the 2 parameters to mindi, 
but that will be done after when a clear line is drawn between
where files are put by each tool.

MONDO_TMP and MINDI_TMP should be reserved for temporary files
MONDO_CACHE and MINDI_CACHE for end results and communication 
between tools. 
These changes were made in order to prepare the USB support in mondo

A first step in the right direction, but still some work to do 
before it's working again.
Testers please be warned.